refactor(controller): 更新企业相关接口权限标识

See merge request s25029301301/onebase-v3-be!566

+++ onebase-module-system/onebase-module-system-core/src/main/java/com/cmsr/onebase/module/system/service/corpapprelation/CorpAppRelationServiceImpl.java
@@ -5,11 +5,10 @@ import com.cmsr.onebase.framework.common.enums.CorpAppReationStatusEnum;
 import com.cmsr.onebase.framework.common.enums.CorpStatusEnum;
 import com.cmsr.onebase.framework.common.pojo.PageResult;
 import com.cmsr.onebase.framework.common.util.object.BeanUtils;
-import com.cmsr.onebase.framework.tenant.core.context.TenantContextHolder;
+import com.cmsr.onebase.framework.security.core.util.SecurityFrameworkUtils;
 import com.cmsr.onebase.module.app.api.app.AppApplicationApi;
 import com.cmsr.onebase.module.app.api.app.dto.ApplicationDTO;
 import com.cmsr.onebase.module.system.dal.database.CorpAppRelationDataRepository;
-import com.cmsr.onebase.module.system.dal.dataobject.corp.CorpDO;
 import com.cmsr.onebase.module.system.dal.dataobject.corpapprelation.CorpAppRelationDO;
 import com.cmsr.onebase.module.system.vo.corp.CorpApplicationRespVO;
 import com.cmsr.onebase.module.system.vo.corpapprelation.*;
@@ -18,18 +17,22 @@ import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.anyline.data.param.ConfigStore;
 import org.anyline.data.param.init.DefaultConfigStore;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static com.cmsr.onebase.framework.common.exception.util.ServiceExceptionUtil.exception;
 import static com.cmsr.onebase.module.system.enums.ErrorCodeConstants.APPLICATION_AUTH_TENANT_NOT_EXISTS;
+import static com.cmsr.onebase.module.system.enums.ErrorCodeConstants.CORP_ID_COMPARE_ERROR;
 
 /**
  * 企业应用关联表 Service 实现类
@@ -47,11 +50,14 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
 
     @Override
     public void createListCorpAppRelation(List<AppAuthTimeReqVO> corpAppRelationInertReqVOList, Long corpId) {
+        if (CollectionUtils.isEmpty(corpAppRelationInertReqVOList)) {
+            return;
+        }
         // 插入
         corpAppRelationInertReqVOList.forEach(corpAppRelationInertReqVO -> {
             // 先删除后插入
             ConfigStore configs = new DefaultConfigStore();
-            configs.eq(CorpAppRelationDO.APPLICATION_ID, corpAppRelationInertReqVO.getId() );
+            configs.eq(CorpAppRelationDO.APPLICATION_ID, corpAppRelationInertReqVO.getId());
             configs.eq(CorpAppRelationDO.CORP_ID, corpId);
             corpAppRelationDataRepository.deleteByConfig(configs);
 
@@ -72,12 +78,12 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
         vo.getApplicationIdList().forEach(appId -> {
             // 先删除后插入
             ConfigStore configs = new DefaultConfigStore();
-            configs.eq(CorpAppRelationDO.APPLICATION_ID,appId );
+            configs.eq(CorpAppRelationDO.APPLICATION_ID, appId);
             configs.eq(CorpAppRelationDO.CORP_ID, vo.getCorpId());
             corpAppRelationDataRepository.deleteByConfig(configs);
 
             // 验证是否重复提交，先删除后插入
-            CorpAppRelationDO corpAppRelationDO =  new CorpAppRelationDO();
+            CorpAppRelationDO corpAppRelationDO = new CorpAppRelationDO();
             corpAppRelationDO.setApplicationId(appId);
             corpAppRelationDO.setAuthorizationTime(vo.getAuthorizationTime());
             corpAppRelationDO.setExpiresTime(vo.getExpiresTime());
@@ -134,6 +140,11 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
 
     @Override
     public PageResult<CorpApplicationRespVO> getCorpAppRelationPage(CorpAppPageReqVO pageReqVO) {
+        Long corpId = pageReqVO.getCorpId();
+        Long loginCorpId = SecurityFrameworkUtils.getLoginUser().getCorpId();
+        if(!Objects.equals(corpId, loginCorpId)){
+            throw exception(CORP_ID_COMPARE_ERROR);
+        }
         Map<Long, ApplicationDTO> applicationMap = getApplicationDoMap(pageReqVO.getAppName());
         List<Long> applicationIds = new ArrayList<>(applicationMap.keySet());
         // 查询原始分页数据
@@ -147,7 +158,6 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
     }
 
 
-
     /**
      * 获取app应用状态描述
      *
@@ -155,17 +165,17 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
      * @return Map<Long, String> key为企业ID，value为企业名称
      */
     public String getCorpStatus(Integer status, LocalDateTime expiresTime) {
-        String statusDesc = "";
+        String showStatus = "";
         if (status != null && status.equals(CorpStatusEnum.DISABLE.getValue())) {
-            statusDesc = CorpAppReationStatusEnum.DISABLE.getName();
+            showStatus = CorpAppReationStatusEnum.DISABLE.getName();
         } else if (expiresTime != null) {
             if (expiresTime.isAfter(java.time.LocalDateTime.now())) {
-                statusDesc = CorpAppReationStatusEnum.ENABLE.getName();
+                showStatus = CorpAppReationStatusEnum.ENABLE.getName();
             } else {
-                statusDesc = CorpAppReationStatusEnum.EXPIRES.getName();
+                showStatus = CorpAppReationStatusEnum.EXPIRES.getName();
             }
         }
-        return statusDesc;
+        return showStatus;
     }
 
     /**
@@ -187,7 +197,7 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
             respVO.setVersionNumber(appDo.getVersionNumber());
             // 获取app应用状态描述
             Integer status = corpDO.getStatus();
-            respVO.setStatusDesc(getCorpStatus(status, corpDO.getExpiresTime()));
+            respVO.setShowStatus(getCorpStatus(status, corpDO.getExpiresTime()));
         }
         return respVO;
     }
@@ -199,7 +209,7 @@ public class CorpAppRelationServiceImpl implements CorpAppRelationService {
 
     @Override
     public List<CorpAppRelationDO> getCorpAppRelationList(CorpAppRelationPageReqVO corpAppRelationPageReqVO) {
-        List<CorpAppRelationDO> corpApplicationRespVOList= corpAppRelationDataRepository.getCorpAppRelationList(corpAppRelationPageReqVO);
+        List<CorpAppRelationDO> corpApplicationRespVOList = corpAppRelationDataRepository.getCorpAppRelationList(corpAppRelationPageReqVO);
         return corpApplicationRespVOList;
     }
 }
